+++ acc_app_optimisation/envs/__init__.py
@@ -17,6 +17,20 @@ if t.TYPE_CHECKING:
     from pyjapc import PyJapc
 
 
+BUILTIN_ENVS = [
+    "cern_awake_env.machine",
+    "cern_awake_env.simulation",
+    "cern_isolde_offline_env",
+    "cern_leir_transfer_line_env",
+    "cern_sps_splitter_opt_env",
+    "cern_sps_tune_env",
+    "cern_sps_zs_alignment_env",
+    "linac3_lebt_tuning",
+    "psb_extr_and_recomb_optim.optimizer",
+    "sps_blowup",
+]
+
+
 class Metadata:
     """Dataclass that reads problem metadata with fallback.